+++ CHANGELOG.md
@@ -260,6 +260,7 @@ Only display line where quantity is greater than 0.
 - Apply shipping coef on WAP price valuation for purchase stock move.
 - Stock move mass invoicing: allow null values after the prompt.
 - Improvement when updating stock move lines from product.
+- Shipping coef: remove duplicated code.
 
 ## Bug Fixes
 - Check duplicate : Apply manual filter before duplicate check and allow to open form view.